backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
networkx==3.2.1
torch==2.5.1
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
faker==22.6.0
google-generativeai==0.8.3
python-multipart==0.0.6
python-dotenv==1.0.1
